Compare results of multiple spflow_models

Compare results of multiple spflow_models

compare_results(
  model_list,
  global_vars = c("model_coherence", "R2_corr", "ll", "AIC", "N_sample"),
  sig_levels = c(`***` = 0.001, `**` = 0.01, `*` = 0.05, `'` = 0.1),
  digits = 3,
  add_dispersion = FALSE
)

Arguments

model_list

a list of models that should be compared (only spflow_model-class() is used)

global_vars

a character indicating which statistics should be reported

sig_levels

a named numeric indicating the codification of significance levels

digits

a numeric indicating to what decimal the results should be rounded

add_dispersion

a logical, if TRUE the standard errors are added in parenthesis

Value

a data.frame

Author

Lukas Dargel

Examples

res_ge  <- spflow(y9 ~ . + P_(DISTANCE), multi_net_usa_ge, "ge_ge")
res_usa <- spflow(y9 ~ . + P_(DISTANCE), multi_net_usa_ge, "usa_usa")
compare_results(list("GE" = res_ge, "US" = res_usa),
                global_vars = c("N_sample", "R2_corr", "model_coherence"))
#>                        GE        US
#> rho_d            0.497***  0.445***
#> rho_o            0.333***  0.312***
#> rho_w           -0.227*** -0.192***
#> (Intercept)     10.199***  8.107***
#> (Intra)          9.871***    8.19**
#> D_X              0.983***  1.009***
#> D_X.lag1         0.509***   0.66***
#> I_X              2.035***  2.036***
#> O_X             -0.759*** -0.787***
#> O_X.lag1        -0.367*** -0.358***
#> P_DISTANCE      -2.622*** -3.056***
#> N_sample              256       781
#> R2_corr             0.992     0.995
#> model_coherence Validated Validated