A function to compute the Jacobean term in the spatial econometric interaction model

These are internal functions called within the estimation procedure

derive_logdet_calculator(
  OW,
  DW,
  n_o,
  n_d,
  model,
  approx_order,
  is_cartesian,
  M_indicator
)

derive_approxldet_cartesian(OW, DW, n_o, n_d, model, approx_order)

derive_approxldet_cartesian2(OW, DW, n_o, n_d, model) # order = 2

derive_approxldet_noncartesian(Wd, Wo, Ww, approx_order, model)

derive_approxldet_noncartesian2(OW, DW, M_indicator, n_o, n_d, model) # order = 2

tracevals2params_cartesian(OW, DW, n_o, n_d, model, approx_order)

tracevals2params_noncartesian(Wd, Wo, Ww, model, approx_order)

trace_template_noncartesian()

tracevals2approxldet(tracevals)

Value

A function that takes the autoregression parameters as an argument and returns the log-determinant value

Details

The Jacobean term of the model corresponds to the log-determinant of the spatial filter matrix, which is costly to compute. To reduce the computational burden of the MLE or MCMC estimators the package uses a power-series approximation of this term. This approximation was first proposed by Martin (1992) , has been adapted to interaction models by LeSage and Pace (2008) and Dargel and Thomas-Agnan (2023) extend the approximation to the non-cartesian and rectangular cases, where the OD-matrix can be sparse and where the list of origins may be different from the list of destinations.

By using this approximation we can avoid to directly calculate the determinant term. Moreover, it is possible to factor out the autoregression parameters from all remaining terms, which means that we do not need to repeat the most costly computations.

References

Dargel L, Thomas-Agnan C (2023). “Efficient Estimation of Spatial Econometric Interaction Models for Sparse OD Matrices.” TSE Working Paper, n. 23-1409, February 2023, https://www.tse-fr.eu/publications/generalized-framework-estimating-spatial-econometric-interaction-models.

LeSage JP, Pace RK (2008). “Spatial Econometric Modeling of Origin-Destination Flows.” Journal of Regional Science, 941--967. https://doi.org/10.1111/j.1467-9787.2008.00573.x.

Martin RJ (1992). “Approximations to the determinant term in gaussian maximum likelihood estimation of some spatial models.” Communications in Statistics - Theory and Methods, 189--205. https://doi.org/10.1080/03610929308831013.

Author

Lukas Dargel