Compute the moment matrices used during the estimation with spflow()

These are internal functions called within the estimation procedure

derive_spflow_moments(spflow_matrices, n_o, n_d, N, wt, na_rm = FALSE)

spflow_moment_var(spflow_matrices, wt, N, n_d, n_o)

spflow_moment_cov(Y, spflow_matrices)

Value

A list containing all moments required for the estimation

Details

The implementation details are derived in Dargel (2021) and Dargel and Thomas-Agnan (2023) .

The variance moment:

The moment (U'U) moment matrix is grouped into 4 x 4 blocks. These 16 blocks are derived as interactions from the four sets of variables U_alpha, U_alpha_I, U_beta, U_gamma. Only ten blocks are unique and the remaining six are inferred by symmetry.

The covariance moment:

The moment (U'y) is grouped into four blocks using the same arguments. Here we can use the same function to compute the inner product of y and its spatial lags with U.

References

Dargel L (2021). “Revisiting estimation methods for spatial econometric interaction models.” Journal of Spatial Econometrics, 10. https://doi.org/10.1007/s43071-021-00016-1.

Dargel L, Thomas-Agnan C (2023). “Efficient Estimation of Spatial Econometric Interaction Models for Sparse OD Matrices.” TSE Working Paper, n. 23-1409, February 2023, https://www.tse-fr.eu/publications/generalized-framework-estimating-spatial-econometric-interaction-models.

Author

Lukas Dargel