Simulated data for stylized versions of Germany and the USA

The package provides two stylized country examples that have been already used byYang et al. (2017) . The first example is a stylized version of Germany with 16 states and second one is based on the USA with 51 states.

There are two sf objects that contain the geography and data for the country examples. The same information is also provided by two spflow_network-class() objects. Additionally, there is an spflow_network_multi-class() that contains the two spflow_network() and four spflow_network_pair(), which contain OD-pairs where each states of each country can be origins or destinations.

For each of the OD-pairs we simulated three flow vectors y1, y2 and y9. The parameters used for this simulation are given as a list: rho contains the auto-regression parameters, delta contains the parameters for the exogenous variables and sd_error is the standard deviation of the Gaussian error term.

multi_net_usa_ge

germany_grid

usa_net

usa_grid

germany_net

simulation_params

Format

An object of class spflow_network_multi of length 1.

An object of class sf (inherits from data.frame) with 16 rows and 3 columns.

An object of class spflow_network of length 1.

An object of class sf (inherits from data.frame) with 51 rows and 3 columns.

An object of class spflow_network of length 1.

An object of class list of length 3.

Source

The stylized versions of Germany and the USA are inspired by https://ialab.it.monash.edu/~dwyer/papers/maptrix.pdf

References

Yang Y, Dwyer T, Goodwin S, Marriott K (2017). “Many-to-Many Geographically-Embedded Flow Visualisation: An Evaluation.” IEEE Transactions on Visualization and Computer Graphics, 411--420. https://doi.org/10.1109/TVCG.2016.2598885.

Author

Lukas Dargel