Create a long form data.frame of origin-destination pairs

The method merges all available information on origins and destinations to the data.frame describing the OD-pairs.

pair_merge(object, ...)

# S4 method for spflow_network_multi
pair_merge(
  object,
  id_net_pair = id(object)[["pairs"]][[1]],
  dest_cols = NULL,
  orig_cols = NULL,
  pair_cols = NULL,
  make_cartesian = FALSE,
  keep_od_keys = TRUE
)

Arguments

object

A spflow_network_multi-class()

...

Arguments to be passed to methods

id_net_pair

A character indicating the id of a spflow_network_pair-class()

dest_cols

A character, indicating the columns to be kept in the final data.frame that contain information on the nodes in the destination network. (TRUE is a shortcut for everything).

orig_cols

A character, indicating the columns to be kept in the final data.frame that contain information on the nodes in the origin network. (TRUE is a shortcut for everything).

pair_cols

A character, indicating the columns to be kept in the final data.frame that contain information on the origin-destination pairs. (TRUE is a shortcut for everything).

make_cartesian

A logical, when set to TRUE the resulting data.frame contains all possible pairs of origins and destination, even if the data in the spflow_network_pair-class() does not have them.

keep_od_keys

A logical, indicating whether the id of the od should be included.

Value

A single data.frame, combining all available information on origins, destinations and OD pairs

See also

Author

Lukas Dargel

Examples

# long form data for flows from Germany to Germany
pair_merge(multi_net_usa_ge,"ge_ge")
#>     ID_DEST ID_ORIG
#> 1        SH      SH
#> 2        HH      SH
#> 3        MV      SH
#> 4        NW      SH
#> 5        HB      SH
#> 6        BB      SH
#> 7        BE      SH
#> 8        RP      SH
#> 9        NI      SH
#> 10       ST      SH
#> 11       SN      SH
#> 12       SL      SH
#> 13       HE      SH
#> 14       TH      SH
#> 15       BW      SH
#> 16       BY      SH
#> 17       SH      HH
#> 18       HH      HH
#> 19       MV      HH
#> 20       NW      HH
#> 21       HB      HH
#> 22       BB      HH
#> 23       BE      HH
#> 24       RP      HH
#> 25       NI      HH
#> 26       ST      HH
#> 27       SN      HH
#> 28       SL      HH
#> 29       HE      HH
#> 30       TH      HH
#> 31       BW      HH
#> 32       BY      HH
#> 33       SH      MV
#> 34       HH      MV
#> 35       MV      MV
#> 36       NW      MV
#> 37       HB      MV
#> 38       BB      MV
#> 39       BE      MV
#> 40       RP      MV
#> 41       NI      MV
#> 42       ST      MV
#> 43       SN      MV
#> 44       SL      MV
#> 45       HE      MV
#> 46       TH      MV
#> 47       BW      MV
#> 48       BY      MV
#> 49       SH      NW
#> 50       HH      NW
#> 51       MV      NW
#> 52       NW      NW
#> 53       HB      NW
#> 54       BB      NW
#> 55       BE      NW
#> 56       RP      NW
#> 57       NI      NW
#> 58       ST      NW
#> 59       SN      NW
#> 60       SL      NW
#> 61       HE      NW
#> 62       TH      NW
#> 63       BW      NW
#> 64       BY      NW
#> 65       SH      HB
#> 66       HH      HB
#> 67       MV      HB
#> 68       NW      HB
#> 69       HB      HB
#> 70       BB      HB
#> 71       BE      HB
#> 72       RP      HB
#> 73       NI      HB
#> 74       ST      HB
#> 75       SN      HB
#> 76       SL      HB
#> 77       HE      HB
#> 78       TH      HB
#> 79       BW      HB
#> 80       BY      HB
#> 81       SH      BB
#> 82       HH      BB
#> 83       MV      BB
#> 84       NW      BB
#> 85       HB      BB
#> 86       BB      BB
#> 87       BE      BB
#> 88       RP      BB
#> 89       NI      BB
#> 90       ST      BB
#> 91       SN      BB
#> 92       SL      BB
#> 93       HE      BB
#> 94       TH      BB
#> 95       BW      BB
#> 96       BY      BB
#> 97       SH      BE
#> 98       HH      BE
#> 99       MV      BE
#> 100      NW      BE
#> 101      HB      BE
#> 102      BB      BE
#> 103      BE      BE
#> 104      RP      BE
#> 105      NI      BE
#> 106      ST      BE
#> 107      SN      BE
#> 108      SL      BE
#> 109      HE      BE
#> 110      TH      BE
#> 111      BW      BE
#> 112      BY      BE
#> 113      SH      RP
#> 114      HH      RP
#> 115      MV      RP
#> 116      NW      RP
#> 117      HB      RP
#> 118      BB      RP
#> 119      BE      RP
#> 120      RP      RP
#> 121      NI      RP
#> 122      ST      RP
#> 123      SN      RP
#> 124      SL      RP
#> 125      HE      RP
#> 126      TH      RP
#> 127      BW      RP
#> 128      BY      RP
#> 129      SH      NI
#> 130      HH      NI
#> 131      MV      NI
#> 132      NW      NI
#> 133      HB      NI
#> 134      BB      NI
#> 135      BE      NI
#> 136      RP      NI
#> 137      NI      NI
#> 138      ST      NI
#> 139      SN      NI
#> 140      SL      NI
#> 141      HE      NI
#> 142      TH      NI
#> 143      BW      NI
#> 144      BY      NI
#> 145      SH      ST
#> 146      HH      ST
#> 147      MV      ST
#> 148      NW      ST
#> 149      HB      ST
#> 150      BB      ST
#> 151      BE      ST
#> 152      RP      ST
#> 153      NI      ST
#> 154      ST      ST
#> 155      SN      ST
#> 156      SL      ST
#> 157      HE      ST
#> 158      TH      ST
#> 159      BW      ST
#> 160      BY      ST
#> 161      SH      SN
#> 162      HH      SN
#> 163      MV      SN
#> 164      NW      SN
#> 165      HB      SN
#> 166      BB      SN
#> 167      BE      SN
#> 168      RP      SN
#> 169      NI      SN
#> 170      ST      SN
#> 171      SN      SN
#> 172      SL      SN
#> 173      HE      SN
#> 174      TH      SN
#> 175      BW      SN
#> 176      BY      SN
#> 177      SH      SL
#> 178      HH      SL
#> 179      MV      SL
#> 180      NW      SL
#> 181      HB      SL
#> 182      BB      SL
#> 183      BE      SL
#> 184      RP      SL
#> 185      NI      SL
#> 186      ST      SL
#> 187      SN      SL
#> 188      SL      SL
#> 189      HE      SL
#> 190      TH      SL
#> 191      BW      SL
#> 192      BY      SL
#> 193      SH      HE
#> 194      HH      HE
#> 195      MV      HE
#> 196      NW      HE
#> 197      HB      HE
#> 198      BB      HE
#> 199      BE      HE
#> 200      RP      HE
#> 201      NI      HE
#> 202      ST      HE
#> 203      SN      HE
#> 204      SL      HE
#> 205      HE      HE
#> 206      TH      HE
#> 207      BW      HE
#> 208      BY      HE
#> 209      SH      TH
#> 210      HH      TH
#> 211      MV      TH
#> 212      NW      TH
#> 213      HB      TH
#> 214      BB      TH
#> 215      BE      TH
#> 216      RP      TH
#> 217      NI      TH
#> 218      ST      TH
#> 219      SN      TH
#> 220      SL      TH
#> 221      HE      TH
#> 222      TH      TH
#> 223      BW      TH
#> 224      BY      TH
#> 225      SH      BW
#> 226      HH      BW
#> 227      MV      BW
#> 228      NW      BW
#> 229      HB      BW
#> 230      BB      BW
#> 231      BE      BW
#> 232      RP      BW
#> 233      NI      BW
#> 234      ST      BW
#> 235      SN      BW
#> 236      SL      BW
#> 237      HE      BW
#> 238      TH      BW
#> 239      BW      BW
#> 240      BY      BW
#> 241      SH      BY
#> 242      HH      BY
#> 243      MV      BY
#> 244      NW      BY
#> 245      HB      BY
#> 246      BB      BY
#> 247      BE      BY
#> 248      RP      BY
#> 249      NI      BY
#> 250      ST      BY
#> 251      SN      BY
#> 252      SL      BY
#> 253      HE      BY
#> 254      TH      BY
#> 255      BW      BY
#> 256      BY      BY

# long form data for flows from Germany to USA
pair_merge(multi_net_usa_ge,"ge_usa")
#>     ID_DEST ID_ORIG
#> 1        AK      SH
#> 2        ME      SH
#> 3        WI      SH
#> 4        VT      SH
#> 5        NH      SH
#> 6        WA      SH
#> 7        ID      SH
#> 8        MT      SH
#> 9        ND      SH
#> 10       MN      SH
#> 11       IL      SH
#> 12       MI      SH
#> 13       NY      SH
#> 14       MA      SH
#> 15       OR      SH
#> 16       NV      SH
#> 17       WY      SH
#> 18       SD      SH
#> 19       IA      SH
#> 20       IN      SH
#> 21       OH      SH
#> 22       PA      SH
#> 23       NJ      SH
#> 24       CT      SH
#> 25       RI      SH
#> 26       CA      SH
#> 27       UT      SH
#> 28       CO      SH
#> 29       NE      SH
#> 30       MO      SH
#> 31       KY      SH
#> 32       WV      SH
#> 33       VA      SH
#> 34       MD      SH
#> 35       DE      SH
#> 36       AZ      SH
#> 37       NM      SH
#> 38       KS      SH
#> 39       AR      SH
#> 40       TN      SH
#> 41       NC      SH
#> 42       SC      SH
#> 43       DC      SH
#> 44       OK      SH
#> 45       LA      SH
#> 46       MS      SH
#> 47       AL      SH
#> 48       GA      SH
#> 49       HI      SH
#> 50       TX      SH
#> 51       FL      SH
#> 52       AK      HH
#> 53       ME      HH
#> 54       WI      HH
#> 55       VT      HH
#> 56       NH      HH
#> 57       WA      HH
#> 58       ID      HH
#> 59       MT      HH
#> 60       ND      HH
#> 61       MN      HH
#> 62       IL      HH
#> 63       MI      HH
#> 64       NY      HH
#> 65       MA      HH
#> 66       OR      HH
#> 67       NV      HH
#> 68       WY      HH
#> 69       SD      HH
#> 70       IA      HH
#> 71       IN      HH
#> 72       OH      HH
#> 73       PA      HH
#> 74       NJ      HH
#> 75       CT      HH
#> 76       RI      HH
#> 77       CA      HH
#> 78       UT      HH
#> 79       CO      HH
#> 80       NE      HH
#> 81       MO      HH
#> 82       KY      HH
#> 83       WV      HH
#> 84       VA      HH
#> 85       MD      HH
#> 86       DE      HH
#> 87       AZ      HH
#> 88       NM      HH
#> 89       KS      HH
#> 90       AR      HH
#> 91       TN      HH
#> 92       NC      HH
#> 93       SC      HH
#> 94       DC      HH
#> 95       OK      HH
#> 96       LA      HH
#> 97       MS      HH
#> 98       AL      HH
#> 99       GA      HH
#> 100      HI      HH
#> 101      TX      HH
#> 102      FL      HH
#> 103      AK      MV
#> 104      ME      MV
#> 105      WI      MV
#> 106      VT      MV
#> 107      NH      MV
#> 108      WA      MV
#> 109      ID      MV
#> 110      MT      MV
#> 111      ND      MV
#> 112      MN      MV
#> 113      IL      MV
#> 114      MI      MV
#> 115      NY      MV
#> 116      MA      MV
#> 117      OR      MV
#> 118      NV      MV
#> 119      WY      MV
#> 120      SD      MV
#> 121      IA      MV
#> 122      IN      MV
#> 123      OH      MV
#> 124      PA      MV
#> 125      NJ      MV
#> 126      CT      MV
#> 127      RI      MV
#> 128      CA      MV
#> 129      UT      MV
#> 130      CO      MV
#> 131      NE      MV
#> 132      MO      MV
#> 133      KY      MV
#> 134      WV      MV
#> 135      VA      MV
#> 136      MD      MV
#> 137      DE      MV
#> 138      AZ      MV
#> 139      NM      MV
#> 140      KS      MV
#> 141      AR      MV
#> 142      TN      MV
#> 143      NC      MV
#> 144      SC      MV
#> 145      DC      MV
#> 146      OK      MV
#> 147      LA      MV
#> 148      MS      MV
#> 149      AL      MV
#> 150      GA      MV
#> 151      HI      MV
#> 152      TX      MV
#> 153      FL      MV
#> 154      AK      NW
#> 155      ME      NW
#> 156      WI      NW
#> 157      VT      NW
#> 158      NH      NW
#> 159      WA      NW
#> 160      ID      NW
#> 161      MT      NW
#> 162      ND      NW
#> 163      MN      NW
#> 164      IL      NW
#> 165      MI      NW
#> 166      NY      NW
#> 167      MA      NW
#> 168      OR      NW
#> 169      NV      NW
#> 170      WY      NW
#> 171      SD      NW
#> 172      IA      NW
#> 173      IN      NW
#> 174      OH      NW
#> 175      PA      NW
#> 176      NJ      NW
#> 177      CT      NW
#> 178      RI      NW
#> 179      CA      NW
#> 180      UT      NW
#> 181      CO      NW
#> 182      NE      NW
#> 183      MO      NW
#> 184      KY      NW
#> 185      WV      NW
#> 186      VA      NW
#> 187      MD      NW
#> 188      DE      NW
#> 189      AZ      NW
#> 190      NM      NW
#> 191      KS      NW
#> 192      AR      NW
#> 193      TN      NW
#> 194      NC      NW
#> 195      SC      NW
#> 196      DC      NW
#> 197      OK      NW
#> 198      LA      NW
#> 199      MS      NW
#> 200      AL      NW
#> 201      GA      NW
#> 202      HI      NW
#> 203      TX      NW
#> 204      FL      NW
#> 205      AK      HB
#> 206      ME      HB
#> 207      WI      HB
#> 208      VT      HB
#> 209      NH      HB
#> 210      WA      HB
#> 211      ID      HB
#> 212      MT      HB
#> 213      ND      HB
#> 214      MN      HB
#> 215      IL      HB
#> 216      MI      HB
#> 217      NY      HB
#> 218      MA      HB
#> 219      OR      HB
#> 220      NV      HB
#> 221      WY      HB
#> 222      SD      HB
#> 223      IA      HB
#> 224      IN      HB
#> 225      OH      HB
#> 226      PA      HB
#> 227      NJ      HB
#> 228      CT      HB
#> 229      RI      HB
#> 230      CA      HB
#> 231      UT      HB
#> 232      CO      HB
#> 233      NE      HB
#> 234      MO      HB
#> 235      KY      HB
#> 236      WV      HB
#> 237      VA      HB
#> 238      MD      HB
#> 239      DE      HB
#> 240      AZ      HB
#> 241      NM      HB
#> 242      KS      HB
#> 243      AR      HB
#> 244      TN      HB
#> 245      NC      HB
#> 246      SC      HB
#> 247      DC      HB
#> 248      OK      HB
#> 249      LA      HB
#> 250      MS      HB
#> 251      AL      HB
#> 252      GA      HB
#> 253      HI      HB
#> 254      TX      HB
#> 255      FL      HB
#> 256      AK      BB
#> 257      ME      BB
#> 258      WI      BB
#> 259      VT      BB
#> 260      NH      BB
#> 261      WA      BB
#> 262      ID      BB
#> 263      MT      BB
#> 264      ND      BB
#> 265      MN      BB
#> 266      IL      BB
#> 267      MI      BB
#> 268      NY      BB
#> 269      MA      BB
#> 270      OR      BB
#> 271      NV      BB
#> 272      WY      BB
#> 273      SD      BB
#> 274      IA      BB
#> 275      IN      BB
#> 276      OH      BB
#> 277      PA      BB
#> 278      NJ      BB
#> 279      CT      BB
#> 280      RI      BB
#> 281      CA      BB
#> 282      UT      BB
#> 283      CO      BB
#> 284      NE      BB
#> 285      MO      BB
#> 286      KY      BB
#> 287      WV      BB
#> 288      VA      BB
#> 289      MD      BB
#> 290      DE      BB
#> 291      AZ      BB
#> 292      NM      BB
#> 293      KS      BB
#> 294      AR      BB
#> 295      TN      BB
#> 296      NC      BB
#> 297      SC      BB
#> 298      DC      BB
#> 299      OK      BB
#> 300      LA      BB
#> 301      MS      BB
#> 302      AL      BB
#> 303      GA      BB
#> 304      HI      BB
#> 305      TX      BB
#> 306      FL      BB
#> 307      AK      BE
#> 308      ME      BE
#> 309      WI      BE
#> 310      VT      BE
#> 311      NH      BE
#> 312      WA      BE
#> 313      ID      BE
#> 314      MT      BE
#> 315      ND      BE
#> 316      MN      BE
#> 317      IL      BE
#> 318      MI      BE
#> 319      NY      BE
#> 320      MA      BE
#> 321      OR      BE
#> 322      NV      BE
#> 323      WY      BE
#> 324      SD      BE
#> 325      IA      BE
#> 326      IN      BE
#> 327      OH      BE
#> 328      PA      BE
#> 329      NJ      BE
#> 330      CT      BE
#> 331      RI      BE
#> 332      CA      BE
#> 333      UT      BE
#> 334      CO      BE
#> 335      NE      BE
#> 336      MO      BE
#> 337      KY      BE
#> 338      WV      BE
#> 339      VA      BE
#> 340      MD      BE
#> 341      DE      BE
#> 342      AZ      BE
#> 343      NM      BE
#> 344      KS      BE
#> 345      AR      BE
#> 346      TN      BE
#> 347      NC      BE
#> 348      SC      BE
#> 349      DC      BE
#> 350      OK      BE
#> 351      LA      BE
#> 352      MS      BE
#> 353      AL      BE
#> 354      GA      BE
#> 355      HI      BE
#> 356      TX      BE
#> 357      FL      BE
#> 358      AK      RP
#> 359      ME      RP
#> 360      WI      RP
#> 361      VT      RP
#> 362      NH      RP
#> 363      WA      RP
#> 364      ID      RP
#> 365      MT      RP
#> 366      ND      RP
#> 367      MN      RP
#> 368      IL      RP
#> 369      MI      RP
#> 370      NY      RP
#> 371      MA      RP
#> 372      OR      RP
#> 373      NV      RP
#> 374      WY      RP
#> 375      SD      RP
#> 376      IA      RP
#> 377      IN      RP
#> 378      OH      RP
#> 379      PA      RP
#> 380      NJ      RP
#> 381      CT      RP
#> 382      RI      RP
#> 383      CA      RP
#> 384      UT      RP
#> 385      CO      RP
#> 386      NE      RP
#> 387      MO      RP
#> 388      KY      RP
#> 389      WV      RP
#> 390      VA      RP
#> 391      MD      RP
#> 392      DE      RP
#> 393      AZ      RP
#> 394      NM      RP
#> 395      KS      RP
#> 396      AR      RP
#> 397      TN      RP
#> 398      NC      RP
#> 399      SC      RP
#> 400      DC      RP
#> 401      OK      RP
#> 402      LA      RP
#> 403      MS      RP
#> 404      AL      RP
#> 405      GA      RP
#> 406      HI      RP
#> 407      TX      RP
#> 408      FL      RP
#> 409      AK      NI
#> 410      ME      NI
#> 411      WI      NI
#> 412      VT      NI
#> 413      NH      NI
#> 414      WA      NI
#> 415      ID      NI
#> 416      MT      NI
#> 417      ND      NI
#> 418      MN      NI
#> 419      IL      NI
#> 420      MI      NI
#> 421      NY      NI
#> 422      MA      NI
#> 423      OR      NI
#> 424      NV      NI
#> 425      WY      NI
#> 426      SD      NI
#> 427      IA      NI
#> 428      IN      NI
#> 429      OH      NI
#> 430      PA      NI
#> 431      NJ      NI
#> 432      CT      NI
#> 433      RI      NI
#> 434      CA      NI
#> 435      UT      NI
#> 436      CO      NI
#> 437      NE      NI
#> 438      MO      NI
#> 439      KY      NI
#> 440      WV      NI
#> 441      VA      NI
#> 442      MD      NI
#> 443      DE      NI
#> 444      AZ      NI
#> 445      NM      NI
#> 446      KS      NI
#> 447      AR      NI
#> 448      TN      NI
#> 449      NC      NI
#> 450      SC      NI
#> 451      DC      NI
#> 452      OK      NI
#> 453      LA      NI
#> 454      MS      NI
#> 455      AL      NI
#> 456      GA      NI
#> 457      HI      NI
#> 458      TX      NI
#> 459      FL      NI
#> 460      AK      ST
#> 461      ME      ST
#> 462      WI      ST
#> 463      VT      ST
#> 464      NH      ST
#> 465      WA      ST
#> 466      ID      ST
#> 467      MT      ST
#> 468      ND      ST
#> 469      MN      ST
#> 470      IL      ST
#> 471      MI      ST
#> 472      NY      ST
#> 473      MA      ST
#> 474      OR      ST
#> 475      NV      ST
#> 476      WY      ST
#> 477      SD      ST
#> 478      IA      ST
#> 479      IN      ST
#> 480      OH      ST
#> 481      PA      ST
#> 482      NJ      ST
#> 483      CT      ST
#> 484      RI      ST
#> 485      CA      ST
#> 486      UT      ST
#> 487      CO      ST
#> 488      NE      ST
#> 489      MO      ST
#> 490      KY      ST
#> 491      WV      ST
#> 492      VA      ST
#> 493      MD      ST
#> 494      DE      ST
#> 495      AZ      ST
#> 496      NM      ST
#> 497      KS      ST
#> 498      AR      ST
#> 499      TN      ST
#> 500      NC      ST
#> 501      SC      ST
#> 502      DC      ST
#> 503      OK      ST
#> 504      LA      ST
#> 505      MS      ST
#> 506      AL      ST
#> 507      GA      ST
#> 508      HI      ST
#> 509      TX      ST
#> 510      FL      ST
#> 511      AK      SN
#> 512      ME      SN
#> 513      WI      SN
#> 514      VT      SN
#> 515      NH      SN
#> 516      WA      SN
#> 517      ID      SN
#> 518      MT      SN
#> 519      ND      SN
#> 520      MN      SN
#> 521      IL      SN
#> 522      MI      SN
#> 523      NY      SN
#> 524      MA      SN
#> 525      OR      SN
#> 526      NV      SN
#> 527      WY      SN
#> 528      SD      SN
#> 529      IA      SN
#> 530      IN      SN
#> 531      OH      SN
#> 532      PA      SN
#> 533      NJ      SN
#> 534      CT      SN
#> 535      RI      SN
#> 536      CA      SN
#> 537      UT      SN
#> 538      CO      SN
#> 539      NE      SN
#> 540      MO      SN
#> 541      KY      SN
#> 542      WV      SN
#> 543      VA      SN
#> 544      MD      SN
#> 545      DE      SN
#> 546      AZ      SN
#> 547      NM      SN
#> 548      KS      SN
#> 549      AR      SN
#> 550      TN      SN
#> 551      NC      SN
#> 552      SC      SN
#> 553      DC      SN
#> 554      OK      SN
#> 555      LA      SN
#> 556      MS      SN
#> 557      AL      SN
#> 558      GA      SN
#> 559      HI      SN
#> 560      TX      SN
#> 561      FL      SN
#> 562      AK      SL
#> 563      ME      SL
#> 564      WI      SL
#> 565      VT      SL
#> 566      NH      SL
#> 567      WA      SL
#> 568      ID      SL
#> 569      MT      SL
#> 570      ND      SL
#> 571      MN      SL
#> 572      IL      SL
#> 573      MI      SL
#> 574      NY      SL
#> 575      MA      SL
#> 576      OR      SL
#> 577      NV      SL
#> 578      WY      SL
#> 579      SD      SL
#> 580      IA      SL
#> 581      IN      SL
#> 582      OH      SL
#> 583      PA      SL
#> 584      NJ      SL
#> 585      CT      SL
#> 586      RI      SL
#> 587      CA      SL
#> 588      UT      SL
#> 589      CO      SL
#> 590      NE      SL
#> 591      MO      SL
#> 592      KY      SL
#> 593      WV      SL
#> 594      VA      SL
#> 595      MD      SL
#> 596      DE      SL
#> 597      AZ      SL
#> 598      NM      SL
#> 599      KS      SL
#> 600      AR      SL
#> 601      TN      SL
#> 602      NC      SL
#> 603      SC      SL
#> 604      DC      SL
#> 605      OK      SL
#> 606      LA      SL
#> 607      MS      SL
#> 608      AL      SL
#> 609      GA      SL
#> 610      HI      SL
#> 611      TX      SL
#> 612      FL      SL
#> 613      AK      HE
#> 614      ME      HE
#> 615      WI      HE
#> 616      VT      HE
#> 617      NH      HE
#> 618      WA      HE
#> 619      ID      HE
#> 620      MT      HE
#> 621      ND      HE
#> 622      MN      HE
#> 623      IL      HE
#> 624      MI      HE
#> 625      NY      HE
#> 626      MA      HE
#> 627      OR      HE
#> 628      NV      HE
#> 629      WY      HE
#> 630      SD      HE
#> 631      IA      HE
#> 632      IN      HE
#> 633      OH      HE
#> 634      PA      HE
#> 635      NJ      HE
#> 636      CT      HE
#> 637      RI      HE
#> 638      CA      HE
#> 639      UT      HE
#> 640      CO      HE
#> 641      NE      HE
#> 642      MO      HE
#> 643      KY      HE
#> 644      WV      HE
#> 645      VA      HE
#> 646      MD      HE
#> 647      DE      HE
#> 648      AZ      HE
#> 649      NM      HE
#> 650      KS      HE
#> 651      AR      HE
#> 652      TN      HE
#> 653      NC      HE
#> 654      SC      HE
#> 655      DC      HE
#> 656      OK      HE
#> 657      LA      HE
#> 658      MS      HE
#> 659      AL      HE
#> 660      GA      HE
#> 661      HI      HE
#> 662      TX      HE
#> 663      FL      HE
#> 664      AK      TH
#> 665      ME      TH
#> 666      WI      TH
#> 667      VT      TH
#> 668      NH      TH
#> 669      WA      TH
#> 670      ID      TH
#> 671      MT      TH
#> 672      ND      TH
#> 673      MN      TH
#> 674      IL      TH
#> 675      MI      TH
#> 676      NY      TH
#> 677      MA      TH
#> 678      OR      TH
#> 679      NV      TH
#> 680      WY      TH
#> 681      SD      TH
#> 682      IA      TH
#> 683      IN      TH
#> 684      OH      TH
#> 685      PA      TH
#> 686      NJ      TH
#> 687      CT      TH
#> 688      RI      TH
#> 689      CA      TH
#> 690      UT      TH
#> 691      CO      TH
#> 692      NE      TH
#> 693      MO      TH
#> 694      KY      TH
#> 695      WV      TH
#> 696      VA      TH
#> 697      MD      TH
#> 698      DE      TH
#> 699      AZ      TH
#> 700      NM      TH
#> 701      KS      TH
#> 702      AR      TH
#> 703      TN      TH
#> 704      NC      TH
#> 705      SC      TH
#> 706      DC      TH
#> 707      OK      TH
#> 708      LA      TH
#> 709      MS      TH
#> 710      AL      TH
#> 711      GA      TH
#> 712      HI      TH
#> 713      TX      TH
#> 714      FL      TH
#> 715      AK      BW
#> 716      ME      BW
#> 717      WI      BW
#> 718      VT      BW
#> 719      NH      BW
#> 720      WA      BW
#> 721      ID      BW
#> 722      MT      BW
#> 723      ND      BW
#> 724      MN      BW
#> 725      IL      BW
#> 726      MI      BW
#> 727      NY      BW
#> 728      MA      BW
#> 729      OR      BW
#> 730      NV      BW
#> 731      WY      BW
#> 732      SD      BW
#> 733      IA      BW
#> 734      IN      BW
#> 735      OH      BW
#> 736      PA      BW
#> 737      NJ      BW
#> 738      CT      BW
#> 739      RI      BW
#> 740      CA      BW
#> 741      UT      BW
#> 742      CO      BW
#> 743      NE      BW
#> 744      MO      BW
#> 745      KY      BW
#> 746      WV      BW
#> 747      VA      BW
#> 748      MD      BW
#> 749      DE      BW
#> 750      AZ      BW
#> 751      NM      BW
#> 752      KS      BW
#> 753      AR      BW
#> 754      TN      BW
#> 755      NC      BW
#> 756      SC      BW
#> 757      DC      BW
#> 758      OK      BW
#> 759      LA      BW
#> 760      MS      BW
#> 761      AL      BW
#> 762      GA      BW
#> 763      HI      BW
#> 764      TX      BW
#> 765      FL      BW
#> 766      AK      BY
#> 767      ME      BY
#> 768      WI      BY
#> 769      VT      BY
#> 770      NH      BY
#> 771      WA      BY
#> 772      ID      BY
#> 773      MT      BY
#> 774      ND      BY
#> 775      MN      BY
#> 776      IL      BY
#> 777      MI      BY
#> 778      NY      BY
#> 779      MA      BY
#> 780      OR      BY
#> 781      NV      BY
#> 782      WY      BY
#> 783      SD      BY
#> 784      IA      BY
#> 785      IN      BY
#> 786      OH      BY
#> 787      PA      BY
#> 788      NJ      BY
#> 789      CT      BY
#> 790      RI      BY
#> 791      CA      BY
#> 792      UT      BY
#> 793      CO      BY
#> 794      NE      BY
#> 795      MO      BY
#> 796      KY      BY
#> 797      WV      BY
#> 798      VA      BY
#> 799      MD      BY
#> 800      DE      BY
#> 801      AZ      BY
#> 802      NM      BY
#> 803      KS      BY
#> 804      AR      BY
#> 805      TN      BY
#> 806      NC      BY
#> 807      SC      BY
#> 808      DC      BY
#> 809      OK      BY
#> 810      LA      BY
#> 811      MS      BY
#> 812      AL      BY
#> 813      GA      BY
#> 814      HI      BY
#> 815      TX      BY
#> 816      FL      BY