Example data for commuting flows within Paris

The package includes an example data set that contains home-to-work commuting flows for 71 municipalities around the center of Paris. The data for the example is stored in three objects

paris10km_municipalities

paris10km_commuteflows

paris10km_neighborhood

Format

An object of class sf (inherits from data.frame) with 71 rows and 6 columns.

An object of class data.frame with 5041 rows and 4 columns.

An object of class list of length 3.

Source

The data combines different public sources (last accessed 2021-05-05).

Three data sets are provided by INSEE. More information on the rights to use and diffuse this data is provided here.

Commuting flows (https://www.insee.fr/fr/statistiques/fichier/3566477/base-texte-flux-mobilite-domicile-lieu-travail-2015.zip)

Number of companies (https://www.insee.fr/fr/statistiques/2021271)

Median income (https://www.insee.fr/fr/statistiques/3560121)

The geographies, population, and area of the municipalities come from OpenDataSoft and are available here. This data set is published under an OPEN LICENCE.

Details

paris10km_municipalities contains information on the municipalities. It contains some socio-economic variables, the identifier of the municipality and a geometry column. The geometry is a MULTIPOLYGON that describes the shape of the region.

paris10km_commuteflows contains information on pairs of municipalities. It is a data.frame with origin and destination identifiers and contains the information on the size of the commuting flows and the distance.

paris10km_neighborhood contains three sparse matrices that represent alternative definitions of the neighborhood of the municipalities. The first is based on contiguity, the second one on distance and the third one is based on the three nearest neighbours.