Create an S4 object that contains information on origin-destination pairs

Create an S4 object that contains information on origin-destination pairs

sp_network_pair(
  orig_net_id,
  dest_net_id,
  pair_data = NULL,
  orig_key_column,
  dest_key_column
)

Arguments

orig_net_id
A character that serves as identifier for the origin network

dest_net_id
A character that serves as identifier for the destination network

pair_data
A data.frame containing information on the origin-destination pairs

orig_key_column
A character indicating the name of the column containing the identifiers of the origins

dest_key_column
A character indicating the name of the column containing the identifiers of the destinations

Value

An S4 class of type sp_network_pair()

See also

Examples

pair_frame <- data.frame(
  ORIG_ID_STATE = rep(germany_grid$ID_STATE, times = 16),
  DEST_ID_STATE = rep(germany_grid$ID_STATE, each = 16))
sp_network_pair("ge","ge",pair_frame,"ORIG_ID_STATE","DEST_ID_STATE")

#> Spatial network pair with id: ge_ge
#> --------------------------------------------------
#> Origin network id: ge (with 16 nodes)
#> Destination network id: ge (with 16 nodes)
#> Number of pairs: 256
#> Completeness of pairs: 100.00% (256/256)
#> 
#> Data on node-pairs:
#>     ORIG_ID_STATE DEST_ID_STATE
#> 1              SH            SH
#> 17             SH            HH
#> 33             SH            MV
#> 49             SH            NW
#> 65             SH            HB
#> 81             SH            BB
#> 97             SH            BE
#> 113            SH            RP
#> 129            SH            NI
#> 145            SH            ST
#> 161            SH            SN
#> 177            SH            SL
#> 193            SH            HE
#> 209            SH            TH
#> 225            SH            BW
#> 241            SH            BY
#> 2              HH            SH
#> 18             HH            HH
#> 34             HH            MV
#> 50             HH            NW
#> 66             HH            HB
#> 82             HH            BB
#> 98             HH            BE
#> 114            HH            RP
#> 130            HH            NI
#> 146            HH            ST
#> 162            HH            SN
#> 178            HH            SL
#> 194            HH            HE
#> 210            HH            TH
#> 226            HH            BW
#> 242            HH            BY
#> 3              MV            SH
#> 19             MV            HH
#> 35             MV            MV
#> 51             MV            NW
#> 67             MV            HB
#> 83             MV            BB
#> 99             MV            BE
#> 115            MV            RP
#> 131            MV            NI
#> 147            MV            ST
#> 163            MV            SN
#> 179            MV            SL
#> 195            MV            HE
#> 211            MV            TH
#> 227            MV            BW
#> 243            MV            BY
#> 4              NW            SH
#> 20             NW            HH
#> 36             NW            MV
#> 52             NW            NW
#> 68             NW            HB
#> 84             NW            BB
#> 100            NW            BE
#> 116            NW            RP
#> 132            NW            NI
#> 148            NW            ST
#> 164            NW            SN
#> 180            NW            SL
#> 196            NW            HE
#> 212            NW            TH
#> 228            NW            BW
#> 244            NW            BY
#> 5              HB            SH
#> 21             HB            HH
#> 37             HB            MV
#> 53             HB            NW
#> 69             HB            HB
#> 85             HB            BB
#> 101            HB            BE
#> 117            HB            RP
#> 133            HB            NI
#> 149            HB            ST
#> 165            HB            SN
#> 181            HB            SL
#> 197            HB            HE
#> 213            HB            TH
#> 229            HB            BW
#> 245            HB            BY
#> 6              BB            SH
#> 22             BB            HH
#> 38             BB            MV
#> 54             BB            NW
#> 70             BB            HB
#> 86             BB            BB
#> 102            BB            BE
#> 118            BB            RP
#> 134            BB            NI
#> 150            BB            ST
#> 166            BB            SN
#> 182            BB            SL
#> 198            BB            HE
#> 214            BB            TH
#> 230            BB            BW
#> 246            BB            BY
#> 7              BE            SH
#> 23             BE            HH
#> 39             BE            MV
#> 55             BE            NW
#> 71             BE            HB
#> 87             BE            BB
#> 103            BE            BE
#> 119            BE            RP
#> 135            BE            NI
#> 151            BE            ST
#> 167            BE            SN
#> 183            BE            SL
#> 199            BE            HE
#> 215            BE            TH
#> 231            BE            BW
#> 247            BE            BY
#> 8              RP            SH
#> 24             RP            HH
#> 40             RP            MV
#> 56             RP            NW
#> 72             RP            HB
#> 88             RP            BB
#> 104            RP            BE
#> 120            RP            RP
#> 136            RP            NI
#> 152            RP            ST
#> 168            RP            SN
#> 184            RP            SL
#> 200            RP            HE
#> 216            RP            TH
#> 232            RP            BW
#> 248            RP            BY
#> 9              NI            SH
#> 25             NI            HH
#> 41             NI            MV
#> 57             NI            NW
#> 73             NI            HB
#> 89             NI            BB
#> 105            NI            BE
#> 121            NI            RP
#> 137            NI            NI
#> 153            NI            ST
#> 169            NI            SN
#> 185            NI            SL
#> 201            NI            HE
#> 217            NI            TH
#> 233            NI            BW
#> 249            NI            BY
#> 10             ST            SH
#> 26             ST            HH
#> 42             ST            MV
#> 58             ST            NW
#> 74             ST            HB
#> 90             ST            BB
#> 106            ST            BE
#> 122            ST            RP
#> 138            ST            NI
#> 154            ST            ST
#> 170            ST            SN
#> 186            ST            SL
#> 202            ST            HE
#> 218            ST            TH
#> 234            ST            BW
#> 250            ST            BY
#> 11             SN            SH
#> 27             SN            HH
#> 43             SN            MV
#> 59             SN            NW
#> 75             SN            HB
#> 91             SN            BB
#> 107            SN            BE
#> 123            SN            RP
#> 139            SN            NI
#> 155            SN            ST
#> 171            SN            SN
#> 187            SN            SL
#> 203            SN            HE
#> 219            SN            TH
#> 235            SN            BW
#> 251            SN            BY
#> 12             SL            SH
#> 28             SL            HH
#> 44             SL            MV
#> 60             SL            NW
#> 76             SL            HB
#> 92             SL            BB
#> 108            SL            BE
#> 124            SL            RP
#> 140            SL            NI
#> 156            SL            ST
#> 172            SL            SN
#> 188            SL            SL
#> 204            SL            HE
#> 220            SL            TH
#> 236            SL            BW
#> 252            SL            BY
#> 13             HE            SH
#> 29             HE            HH
#> 45             HE            MV
#> 61             HE            NW
#> 77             HE            HB
#> 93             HE            BB
#> 109            HE            BE
#> 125            HE            RP
#> 141            HE            NI
#> 157            HE            ST
#> 173            HE            SN
#> 189            HE            SL
#> 205            HE            HE
#> 221            HE            TH
#> 237            HE            BW
#> 253            HE            BY
#> 14             TH            SH
#> 30             TH            HH
#> 46             TH            MV
#> 62             TH            NW
#> 78             TH            HB
#> 94             TH            BB
#> 110            TH            BE
#> 126            TH            RP
#> 142            TH            NI
#> 158            TH            ST
#> 174            TH            SN
#> 190            TH            SL
#> 206            TH            HE
#> 222            TH            TH
#> 238            TH            BW
#> 254            TH            BY
#> 15             BW            SH
#> 31             BW            HH
#> 47             BW            MV
#> 63             BW            NW
#> 79             BW            HB
#> 95             BW            BB
#> 111            BW            BE
#> 127            BW            RP
#> 143            BW            NI
#> 159            BW            ST
#> 175            BW            SN
#> 191            BW            SL
#> 207            BW            HE
#> 223            BW            TH
#> 239            BW            BW
#> 255            BW            BY
#> 16             BY            SH
#> 32             BY            HH
#> 48             BY            MV
#> 64             BY            NW
#> 80             BY            HB
#> 96             BY            BB
#> 112            BY            BE
#> 128            BY            RP
#> 144            BY            NI
#> 160            BY            ST
#> 176            BY            SN
#> 192            BY            SL
#> 208            BY            HE
#> 224            BY            TH
#> 240            BY            BW
#> 256            BY            BY
#>