Class sp_network_pair

An S4 class which holds information on origin-destination (OD) pairs. Each OD pair is composed of two nodes (see sp_network_nodes-class()). All origins belong to the same (origin-) network and all destination belong to the same (destination-) network. It is possible to choose the same network for origins and destinations, which enables to represent OD pairs within the same network.

# S4 method for sp_network_pair
dat(object)

# S4 method for sp_network_pair
dat(object) <- value

# S4 method for sp_network_pair
id(object)

# S4 method for sp_network_pair
id(object) <- value

# S4 method for sp_network_pair
npairs(object)

# S4 method for sp_network_pair
nnodes(object)

Arguments

object
A sp_network_pair-class

value
An object to replace the existing id/data

Slots

orig_net_id

A character that serves as identifier for the origin network

orig_nnodes

A numeric that represents the number of nodes in the origin network

dest_net_id

A character that serves as identifier for the destination network

dest_nnodes

A numeric that represents the number of nodes in the destination network

network_pair_id

A character identifying the pair of networks

pair_data

A data.frame containing information on origin-destination pairs

npairs

A numeric indicating the number of origin-destination pairs

See also

Examples

## access the data describing the node pairs

net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
dat(net_pair_ge_ge)

#>     ID_DEST ID_ORIG DISTANCE           y9          y2          y1
#> 1        SH      SH 0.000000  65.47666194  62.4441803  44.0149303
#> 2        HH      SH 1.000000  49.81614669  38.4218761  19.2511950
#> 3        MV      SH 1.414214  62.82043453  41.7229297  21.7244876
#> 4        NW      SH 2.236068  25.26122242  21.0210762   7.0721976
#> 5        HB      SH 2.000000  47.08688938  36.3716105  21.6221087
#> 6        BB      SH 2.236068  59.29729755  41.9998462  27.4832586
#> 7        BE      SH 2.828427  39.57834501  31.7941562  13.4353579
#> 8        RP      SH 3.162278  17.04525619  21.1416753   6.9845493
#> 9        NI      SH 3.000000  46.87695957  42.2433642  28.7095345
#> 10       ST      SH 3.162278  38.15496191  34.9979338  18.4394640
#> 11       SN      SH 3.605551  32.06700033  25.1471007  10.2865048
#> 12       SL      SH 4.123106  13.02807096   9.4248620   3.3787928
#> 13       HE      SH 4.000000  18.04454600  15.2514303  10.2864399
#> 14       TH      SH 4.123106  13.09297357  13.2530098   5.1307710
#> 15       BW      SH 5.000000  -4.87525671   1.4278292  -2.2266484
#> 16       BY      SH 5.099020  -1.42195628   1.5178378  -4.6188628
#> 17       SH      HH 1.000000  40.55316590  36.8162078  11.7605502
#> 18       HH      HH 0.000000  72.47965966  67.1506073  52.2430862
#> 19       MV      HH 1.000000  53.00094688  41.7720101  21.8507413
#> 20       NW      HH 1.414214  27.23380420  24.1734129   6.4346922
#> 21       HB      HH 1.000000  47.81606194  38.3443205  18.4845833
#> 22       BB      HH 1.414214  57.83247244  43.1291155  26.0247738
#> 23       BE      HH 2.236068  32.88483516  30.4789047  17.6234985
#> 24       RP      HH 2.236068  17.57598851  17.4857793   8.1736247
#> 25       NI      HH 2.000000  46.95993964  33.6341196  28.5455068
#> 26       ST      HH 2.236068  34.75193368  31.9505241  18.5923647
#> 27       SN      HH 2.828427  28.86674326  24.3996847   9.2376043
#> 28       SL      HH 3.162278  11.62703537  10.0416705   2.8486853
#> 29       HE      HH 3.000000  14.34609586  14.5246242   6.9506466
#> 30       TH      HH 3.162278  13.25845452  11.5340153   2.4157675
#> 31       BW      HH 4.000000  -4.27962637  -0.5795747  -3.8516654
#> 32       BY      HH 4.123106  -6.34905483   1.5742352  -7.9240439
#> 33       SH      MV 1.414214  36.77733424  29.8372610   7.1900916
#> 34       HH      MV 1.000000  38.84794364  28.9932027  12.7683050
#> 35       MV      MV 0.000000  90.30949870  81.3090303  61.9027553
#> 36       NW      MV 2.236068  14.70685202  14.3416449   2.0207801
#> 37       HB      MV 1.414214  42.82514816  32.5902495  20.7035269
#> 38       BB      MV 1.000000  54.21806708  37.6670576  21.6569976
#> 39       BE      MV 1.414214  37.52690014  33.9373686  13.6471678
#> 40       RP      MV 2.828427   8.40050849   8.5195730  -1.0096851
#> 41       NI      MV 2.236068  38.42565484  32.0435245  23.4882189
#> 42       ST      MV 2.000000  31.85181053  22.0071844  14.8425066
#> 43       SN      MV 2.236068  21.06564311  18.7948955  11.2956269
#> 44       SL      MV 3.605551   5.85911689   3.8657630  -5.2345953
#> 45       HE      MV 3.162278  10.23933051   4.1749580   2.6677141
#> 46       TH      MV 3.000000   2.15528966   2.0428981  -1.4538504
#> 47       BW      MV 4.123106 -13.84623429 -12.0039429 -10.2638466
#> 48       BY      MV 4.000000 -14.10405517 -12.3729396 -10.6591575
#> 49       SH      NW 2.236068  33.81668692  29.1782902  11.0241089
#> 50       HH      NW 1.414214  47.39555217  40.1458691  20.6548041
#> 51       MV      NW 2.236068  49.30487897  37.0133005  20.2166346
#> 52       NW      NW 0.000000  58.11616664  55.3284768  38.4930619
#> 53       HB      NW 1.000000  60.01916018  48.9540228  24.4133011
#> 54       BB      NW 2.000000  62.23956284  51.5942592  29.6735869
#> 55       BE      NW 3.000000  39.00276845  37.3489529  17.9659786
#> 56       RP      NW 1.000000  42.13352440  35.9117821  16.9134707
#> 57       NI      NW 1.414214  65.60831246  47.2506903  32.8288660
#> 58       ST      NW 2.236068  49.30936468  39.1216416  23.4936572
#> 59       SN      NW 3.162278  35.92298379  30.9072514  16.8356953
#> 60       SL      NW 2.000000  28.80588856  27.9101539   9.2870682
#> 61       HE      NW 2.236068  32.70505498  27.8703510  13.6854595
#> 62       TH      NW 2.828427  22.74700828  20.3633893  10.4555233
#> 63       BW      NW 3.162278   8.80023419  16.1058974   1.8236031
#> 64       BY      NW 3.605551   5.22300252  12.9570744   4.2233542
#> 65       SH      HB 2.000000  16.56540045  16.8878970   4.5893649
#> 66       HH      HB 1.000000  35.13547602  29.7601277  10.7271906
#> 67       MV      HB 1.414214  42.25458072  28.7740024  11.6847018
#> 68       NW      HB 1.000000  25.54051747  23.9481983   4.0450576
#> 69       HB      HB 0.000000  89.02182365  76.7185155  63.5528825
#> 70       BB      HB 1.000000  56.41819875  39.6830907  22.1095927
#> 71       BE      HB 2.000000  27.76474675  23.0763080   7.3797387
#> 72       RP      HB 1.414214  18.03045733  19.7176256   5.3375325
#> 73       NI      HB 1.000000  53.75087977  39.4366248  22.4399443
#> 74       ST      HB 1.414214  38.34417155  29.9034015  15.1906296
#> 75       SN      HB 2.236068  21.99194832  21.2403477   8.2975049
#> 76       SL      HB 2.236068  11.58138286   6.3210010  -0.2323784
#> 77       HE      HB 2.000000  14.70118032  10.3699686   6.9659828
#> 78       TH      HB 2.236068   6.57253370  11.0756643  -1.1983676
#> 79       BW      HB 3.000000  -7.23978905  -4.0734118 -12.5766922
#> 80       BY      HB 3.162278  -9.42286864  -8.8425883  -6.4140399
#> 81       SH      BB 2.236068   9.82400654   5.7898006  -0.2414100
#> 82       HH      BB 1.414214  25.40023776  18.1608225   4.3113828
#> 83       MV      BB 1.000000  35.53036026  28.2408003  12.4939668
#> 84       NW      BB 2.000000   7.76381918   4.4703695  -5.7761188
#> 85       HB      BB 1.000000  35.03820873  25.5465492  11.9171314
#> 86       BB      BB 0.000000  99.34930013  87.9448237  71.1883918
#> 87       BE      BB 1.000000  36.41369407  29.4860840  10.6480761
#> 88       RP      BB 2.236068   4.85739469   2.2057855   1.1373321
#> 89       NI      BB 1.414214  40.98940547  26.4283691  20.9660981
#> 90       ST      BB 1.000000  37.72715089  21.1872951  12.3738775
#> 91       SN      BB 1.414214  30.31287362  21.3671705   5.4764129
#> 92       SL      BB 2.828427  -0.90707784  -4.2314126  -7.9317023
#> 93       HE      BB 2.236068   6.14797035  -0.2556348   2.2154239
#> 94       TH      BB 2.000000   2.28194760   0.2746128  -2.7357541
#> 95       BW      BB 3.162278 -14.01824418 -12.2731459 -13.2146045
#> 96       BY      BB 3.000000 -17.19934046 -17.1379079 -14.1383373
#> 97       SH      BE 2.828427  22.72536954  15.3131977   3.4151936
#> 98       HH      BE 2.236068  29.50817165  22.2623094   8.4336738
#> 99       MV      BE 1.414214  44.52806175  31.7511912  17.1632983
#> 100      NW      BE 3.000000  10.41525110  13.4549148  -1.1852000
#> 101      HB      BE 2.000000  35.73880459  30.1794545  20.2608933
#> 102      BB      BE 1.000000  58.31027183  41.4927838  25.5394235
#> 103      BE      BE 0.000000  75.66111651  63.8661108  55.1498902
#> 104      RP      BE 3.162278   9.46669307  16.6579313  -2.8502323
#> 105      NI      BE 2.236068  41.79850572  33.4523845  23.1604004
#> 106      ST      BE 1.414214  45.50924362  34.9199567  17.5492519
#> 107      SN      BE 1.000000  38.97716487  32.2619102  12.0541476
#> 108      SL      BE 3.605551   2.34746947   9.7654246   3.5765645
#> 109      HE      BE 2.828427  14.07312699  14.9461331   3.7412976
#> 110      TH      BE 2.236068  15.77488217  12.7378352   5.3285766
#> 111      BW      BE 3.605551  -7.04878956  -0.7654981  -3.8237813
#> 112      BY      BE 3.162278  -3.44044031  -1.9398801  -3.7304526
#> 113      SH      RP 3.162278  21.10315147  26.1859478   6.7026561
#> 114      HH      RP 2.236068  36.19095121  31.1484029  17.4541277
#> 115      MV      RP 2.828427  42.27536094  40.0468626  19.4958948
#> 116      NW      RP 1.000000  38.55090973  29.3906317  10.5630042
#> 117      HB      RP 1.414214  55.28426669  41.8235888  21.3069755
#> 118      BB      RP 2.236068  54.06926065  45.8716747  31.4766595
#> 119      BE      RP 3.162278  33.81048319  33.3992447  13.5961607
#> 120      RP      RP 0.000000  62.17545924  57.0616494  38.6488567
#> 121      NI      RP 1.000000  68.94153510  47.7683989  34.2815187
#> 122      ST      RP 2.000000  46.20820160  39.1111084  23.2033177
#> 123      SN      RP 3.000000  32.54721873  29.8664658  11.2323313
#> 124      SL      RP 1.000000  36.31895178  27.0275719   9.8540872
#> 125      HE      RP 1.414214  42.59432566  30.4410478  16.0218971
#> 126      TH      RP 2.236068  24.72175986  23.4409110  12.4105887
#> 127      BW      RP 2.236068  14.97140248  14.6678765   5.3040441
#> 128      BY      RP 2.828427  11.55632758  10.5271598   1.3688438
#> 129      SH      NI 3.000000  -2.33003402  -0.3906335  -8.0183765
#> 130      HH      NI 2.000000  12.33215335   9.8493107   1.5321335
#> 131      MV      NI 2.236068  21.14881769  12.4044666   5.9565028
#> 132      NW      NI 1.414214  16.12401310   8.8038553  -2.5832172
#> 133      HB      NI 1.000000  33.87171930  20.6822806   9.2441184
#> 134      BB      NI 1.414214  38.98072656  26.3019252  16.2774960
#> 135      BE      NI 2.236068  12.65037330   8.5593139   0.7801730
#> 136      RP      NI 1.000000  19.04880689   7.5920950  -1.6710894
#> 137      NI      NI 0.000000 104.93081734  94.4778511  88.0191290
#> 138      ST      NI 1.000000  34.39154857  19.7371639  12.2728004
#> 139      SN      NI 2.000000  18.01604574   7.4374141  -2.2270094
#> 140      SL      NI 1.414214  10.14885141  11.3458239  -2.5782599
#> 141      HE      NI 1.000000  18.15933879   8.2167184   3.4247009
#> 142      TH      NI 1.414214   9.07381841   7.0419051  -5.5942893
#> 143      BW      NI 2.000000 -13.01774744 -11.3474341  -9.5887488
#> 144      BY      NI 2.236068 -14.76511579 -11.2806103 -12.8145525
#> 145      SH      ST 3.162278   4.97569872   7.3969096   1.6251016
#> 146      HH      ST 2.236068  19.17116149  13.6487476   7.7801887
#> 147      MV      ST 2.000000  27.43973632  21.7264794  14.3095332
#> 148      NW      ST 2.236068  14.40049327  10.3359500   0.3546633
#> 149      HB      ST 1.414214  43.30186280  29.7091353  16.9417990
#> 150      BB      ST 1.000000  49.22013611  33.9740345  24.9038862
#> 151      BE      ST 1.414214  38.25055369  28.8772547  15.2612839
#> 152      RP      ST 2.000000  15.87250558  13.0979802   2.3546385
#> 153      NI      ST 1.000000  52.52228773  34.9255453  19.8483264
#> 154      ST      ST 0.000000  84.76709380  76.8972366  64.2059092
#> 155      SN      ST 1.000000  35.80929537  29.6064746  11.4825510
#> 156      SL      ST 2.236068   9.76098493   8.9787982   0.8525339
#> 157      HE      ST 1.414214  23.99782992  17.9810016   8.5225080
#> 158      TH      ST 1.000000  20.22635545  16.2567086   6.9927845
#> 159      BW      ST 2.236068   0.01413962  -1.2942735  -5.8727402
#> 160      BY      ST 2.000000  -1.40519041  -4.3733538  -4.6167621
#> 161      SH      SN 3.605551  15.96027924  14.5665395   3.4790306
#> 162      HH      SN 2.828427  23.42917990  22.4052044   8.2542711
#> 163      MV      SN 2.236068  34.16342825  31.2351306  16.6743308
#> 164      NW      SN 3.162278  20.20734598  12.7864957  -2.1602482
#> 165      HB      SN 2.236068  37.76175525  33.5205080  14.9377109
#> 166      BB      SN 1.414214  58.73388679  43.3348418  22.6732769
#> 167      BE      SN 1.000000  48.04843577  38.0899818  15.2927526
#> 168      RP      SN 3.000000  14.86591136  16.6495985   2.4533909
#> 169      NI      SN 2.000000  50.16189110  35.2814483  26.5396103
#> 170      ST      SN 1.000000  51.86829875  37.1693759  23.1070831
#> 171      SN      SN 0.000000  75.82007985  73.4815414  51.7952329
#> 172      SL      SN 3.162278  11.21727499  11.4845542   0.8194891
#> 173      HE      SN 2.236068  21.73970968  20.9832705   7.2655612
#> 174      TH      SN 1.414214  30.02034693  18.6203466   7.3281819
#> 175      BW      SN 2.828427   1.11485934   3.9668280  -2.6633395
#> 176      BY      SN 2.236068   2.54438417   6.9842106   0.6590311
#> 177      SH      SL 4.123106  20.05419178  19.1726753   5.2880334
#> 178      HH      SL 3.162278  32.21425788  25.2498309   8.9961885
#> 179      MV      SL 3.605551  38.86548208  30.9364585  17.0004776
#> 180      NW      SL 2.000000  28.98132423  24.3100328  12.4243736
#> 181      HB      SL 2.236068  45.61684833  39.0916234  24.3607129
#> 182      BB      SL 2.828427  55.00326595  43.2440588  24.3082208
#> 183      BE      SL 3.605551  34.88398461  25.6261874  10.8877931
#> 184      RP      SL 1.000000  37.92423207  35.0759800  19.5561210
#> 185      NI      SL 1.414214  68.24443094  46.4925575  32.0075154
#> 186      ST      SL 2.236068  47.87762470  38.7376254  22.0397507
#> 187      SN      SL 3.162278  36.85341445  30.1291164  14.6543154
#> 188      SL      SL 0.000000  62.77560397  57.4225829  39.2219211
#> 189      HE      SL 1.000000  45.08061346  32.2897933  16.4007046
#> 190      TH      SL 2.000000  28.84333281  25.6911871   8.1191152
#> 191      BW      SL 1.414214  24.71112855  22.7848919   5.4711701
#> 192      BY      SL 2.236068  16.24591593  18.3732825   4.7294949
#> 193      SH      HE 4.000000  14.55622134  10.3108547   0.3991525
#> 194      HH      HE 3.000000  25.52250055  19.6959466  11.6485368
#> 195      MV      HE 3.162278  31.99621588  30.4683077  15.9204265
#> 196      NW      HE 2.236068  28.42637340  22.4783643   4.3619374
#> 197      HB      HE 2.000000  45.34768050  34.7828657  19.9669356
#> 198      BB      HE 2.236068  48.58190803  40.8499821  25.9879584
#> 199      BE      HE 2.828427  32.93967272  30.6466243  15.2654806
#> 200      RP      HE 1.414214  32.50218148  23.2190639   9.8002342
#> 201      NI      HE 1.000000  68.78519913  48.6553604  30.2162978
#> 202      ST      HE 1.414214  53.90243520  38.6018958  21.3553596
#> 203      SN      HE 2.236068  34.35766191  29.3596759  16.3032910
#> 204      SL      HE 1.000000  35.87066818  28.6918847  13.1787152
#> 205      HE      HE 0.000000  74.60652320  68.8912887  55.3259460
#> 206      TH      HE 1.000000  33.31578247  27.4843842  10.4538898
#> 207      BW      HE 1.000000  22.28161102  22.3776547   5.6750543
#> 208      BY      HE 1.414214  21.48989071  20.8175599   4.5925114
#> 209      SH      TH 4.123106  23.61202094  18.8044444   3.7806699
#> 210      HH      TH 3.162278  31.79422806  30.7764794   9.6298913
#> 211      MV      TH 3.000000  41.02020882  36.4501976  20.2169934
#> 212      NW      TH 2.828427  27.04592487  27.0119802   4.4394309
#> 213      HB      TH 2.236068  49.90014294  38.8458541  21.9142111
#> 214      BB      TH 2.000000  56.19281899  48.5537240  28.8052575
#> 215      BE      TH 2.236068  40.71172119  41.1868784  19.1528642
#> 216      RP      TH 2.236068  33.01454784  26.2166268  11.2439819
#> 217      NI      TH 1.414214  65.53425849  46.9646292  30.7999850
#> 218      ST      TH 1.000000  56.48890510  45.4328465  27.8489222
#> 219      SN      TH 1.414214  45.37440296  40.3268847  22.4679602
#> 220      SL      TH 2.000000  27.05831228  26.5084856  12.7956910
#> 221      HE      TH 1.000000  41.77114417  37.4964020  13.9031744
#> 222      TH      TH 0.000000  65.02277171  58.2867694  45.2324857
#> 223      BW      TH 1.414214  27.12303322  24.1827087   6.6785808
#> 224      BY      TH 1.000000  28.69811093  25.6728472   9.1831619
#> 225      SH      BW 5.000000  23.76135159  24.0218972   8.9440665
#> 226      HH      BW 4.000000  35.81098586  32.5559219  14.8466279
#> 227      MV      BW 4.123106  42.80929494  39.8040595  19.8801567
#> 228      NW      BW 3.162278  31.08793804  29.9299475  13.0014812
#> 229      HB      BW 3.000000  53.14295803  44.0028422  25.5089181
#> 230      BB      BW 3.162278  57.90691688  49.8835486  31.2651201
#> 231      BE      BW 3.605551  47.30633174  39.7240896  23.1491996
#> 232      RP      BW 2.236068  38.35320271  33.5132775  16.6606278
#> 233      NI      BW 2.000000  68.67827811  56.3647914  35.3897623
#> 234      ST      BW 2.236068  55.88621876  45.5931273  22.2133870
#> 235      SN      BW 2.828427  44.37631798  38.1521256  24.5981466
#> 236      SL      BW 1.414214  42.27315149  40.8613632  16.2920057
#> 237      HE      BW 1.000000  52.00013068  43.6218389  27.0066123
#> 238      TH      BW 1.414214  46.10569417  35.0123484  18.1066464
#> 239      BW      BW 0.000000  57.05339319  51.3525661  39.4268604
#> 240      BY      BW 1.000000  35.35497288  29.3818549  12.3805794
#> 241      SH      BY 5.099020  23.48026573  22.9117781   8.3592163
#> 242      HH      BY 4.123106  34.66069689  30.4549930  15.0319324
#> 243      MV      BY 4.000000  47.24499967  37.1530465  19.9610919
#> 244      NW      BY 3.605551  27.64091265  27.3944964   9.3258919
#> 245      HB      BY 3.162278  50.97562332  39.3652372  25.5578833
#> 246      BB      BY 3.000000  63.80138354  45.6342727  29.3018456
#> 247      BE      BY 3.162278  46.90454157  40.4279836  15.5711354
#> 248      RP      BY 2.828427  32.68443071  31.2099390  13.0962735
#> 249      NI      BY 2.236068  65.79041682  47.6231772  35.2797106
#> 250      ST      BY 2.000000  59.30153936  45.5470927  28.5091019
#> 251      SN      BY 2.236068  47.54955681  38.1359274  22.1582650
#> 252      SL      BY 2.236068  35.89764401  27.1525985  11.3409264
#> 253      HE      BY 1.414214  50.78431771  38.6840958  21.6853196
#> 254      TH      BY 1.000000  48.11454656  37.3034835  20.4690518
#> 255      BW      BY 1.000000  39.66988281  29.7625322  14.9864907
#> 256      BY      BY 0.000000  56.28289716  52.6281551  35.8834108

## access the id of a network pair

net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
id(net_pair_ge_ge)

#>    pair    orig    dest 
#> "ge_ge"    "ge"    "ge" 

id(net_pair_ge_ge) <- "Germany_Germany"

## access the number of node pairs in a network pair

net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
npairs(net_pair_ge_ge)

#> [1] 256

## access the number of origin and destination nodes in a network pair
net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
nnodes(net_pair_ge_ge)

#> orig dest 
#>   16   16 

nnodes(net_pair_ge_ge)["orig"]

#> orig 
#>   16 

nnodes(net_pair_ge_ge)["dest"]

#> dest 
#>   16 

prod(nnodes(net_pair_ge_ge) == npairs(net_pair_ge_ge))

#> [1] 0