Generic functions for the spflow_network_classes() and the spflow_model-class()

Most of these generic functions link to simple accessing or replacement methods, and do not have their own documentation page. The associated methods are documented within the corresponding classes.

Generics with more complex methods have their dedicated documentation pages.

Also note that some generic functions are currently not exported and meant to be used by package developers only.

complete_pairs(object, ...)

coord(object, ...) # internal

dat(object, ...)

dat(object, ...) <- value

id(object, ...)

id(object, ...) <- value

impacts_matrix(object, ...)

mcmc_results(object)

neighborhood(object, ...)

neighborhood(object) <- value

nnodes(object, ...)

npairs(object, ...)

pull_member(object, ...)

results(object, ...)

results(object) <- value # internal

results_flat(object, ...)

sd_error(object)

update_dat(object, ...)

varcov(object)

Arguments

object

A spflow_model-class() or spflow_network_classes()

...

Arguments passed on to methods

value

An appropriate replacement value

See also