Class spflow_model

An S4 class that contains the estimation results of spatial econometric interaction models estimated by the spflow() function.

There are four subclasses that are specific to the chosen estimation method (OLS, MLE, Bayesian MCMC or S2SLS). They contain some additional information specific to the corresponding method but most behaviours and data are identical among them.

# S4 method for spflow_model
coef(object)

# S4 method for spflow_model
fitted(object)

# S4 method for spflow_model
nobs(object)

# S4 method for spflow_model
predict(object, ..., type = "BP")

# S4 method for spflow_model
resid(object)

# S4 method for spflow_model
results(object)

# S4 method for spflow_model
sd_error(object)

# S4 method for spflow_model_mle
logLik(object)

# S4 method for spflow_model_mle_s2sls_ols
varcov(object)

# S4 method for spflow_model_mcmc
mcmc_results(object)

Arguments

object
A spflow_model-class()

...
Further arguments passed to the prediction function

type
A character declaring the type of prediction (for now only "BP")

Slots

estimation_results

A data.frame that contains the main results() of the estimation

estimation_control

A list that contains all control parameters of the estimation (see spflow_control())

N

A numeric that corresponds to the number of origin-destination pairs (the sample size in this model)

sd_error

A numeric representing the standard deviation of the residual

R2_corr

A numeric that serves as a goodness of fit criterion. The R2_corr is computed as squared correlation between the fitted values and the observed values of the dependent variable.

resid

A numeric vector of regression residuals

fitted

A numeric vector of fitted values computed as the in sample prediction trend signal (TS) prediction described by @Goulard2017

spatial_filter_matrix

A matrix (can be sparse) or NULL

design_matrix

A matrix (can be sparse) or NULL

varcov

A matrix or NULL

ll

A numeric or NULL

AIC

A numeric or NULL

BIC

A numeric or NULL

mcmc_results

A data.frame containing the estimated parameters for each iteration of the MCMC sampling procedure

Main results

The main results are accessed with the results() method. They are given in the form of a data frame with the following columns;

est: value of the estimated parameter

sd: value of the standard deviation of the parameter

t.test: value of the t-statistic under the two-sided hypothesis that the parameter value is 0.

p.val: the p-value associated to the t-test

quant_025: for Bayesian estimation the lower bound of 95% interval

quant_975: for Bayesian estimation the upper bound of 95% interval

See also

Examples

spflow_results <- spflow(y9 ~ . + G_(DISTANCE),multi_net_usa_ge)

# General methods
results(spflow_results) # data.frame of main results

#>                     est         sd     t.stat    p.value
#> rho_d         0.4756279 0.03020574  15.746275 0.02018782
#> rho_o         0.3574141 0.03375372  10.588883 0.02997187
#> rho_w        -0.2545629 0.04280474  -5.947073 0.05302773
#> (Intercept)   9.9978953 1.91776076   5.213317 0.06032435
#> (Intra)      11.3226073 3.06535087   3.693739 0.08415813
#> DEST_X        0.9356702 0.06252175  14.965516 0.02123798
#> DEST_X.lag1   0.6227750 0.10745954   5.795437 0.05438867
#> ORIG_X       -0.7762048 0.03661436 -21.199462 0.01500388
#> ORIG_X.lag1  -0.3235546 0.08235261  -3.928893 0.07933319
#> INTRA_X       1.9521538 0.08140230  23.981555 0.01326543
#> INTRA_X.lag1 -0.3398274 0.19359618  -1.755341 0.16483179
#> DISTANCE     -2.7980934 0.33924072  -8.248106 0.03840444

coef(spflow_results) # vector of estimated coefficients

#>        rho_d        rho_o        rho_w  (Intercept)      (Intra)       DEST_X 
#>    0.4756279    0.3574141   -0.2545629    9.9978953   11.3226073    0.9356702 
#>  DEST_X.lag1       ORIG_X  ORIG_X.lag1      INTRA_X INTRA_X.lag1     DISTANCE 
#>    0.6227750   -0.7762048   -0.3235546    1.9521538   -0.3398274   -2.7980934 

fitted(spflow_results) # vector of fitted values

#>   [1]  66.1267828  50.8021927  60.6250416  25.1916726  46.8565155  56.0118971
#>   [7]  38.7696996  19.6152165  48.0712706  39.1508026  29.7767595  12.5890118
#>  [13]  17.3229782  13.1320476  -3.6630774  -4.9207151  39.8219040  75.9108164
#>  [19]  51.8547878  28.4200935  50.2001187  58.7068882  34.8407943  18.8490466
#>  [25]  47.8348970  38.0125079  27.0708701  11.3299510  16.4816101  10.6645990
#>  [31]  -5.2847140  -5.8920093  35.6097554  37.3716769  88.0103271  13.3712151
#>  [37]  42.0779969  54.8250795  38.5008610   9.0883261  39.4873983  32.1383351
#>  [43]  23.4416381   1.5620559   7.6516709   4.3447435 -13.1627405 -13.3327655
#>  [49]  32.2566333  47.9377102  48.9540515  58.2177857  60.6451912  59.6685555
#>  [55]  39.6811331  39.2702603  69.2748377  48.2417675  35.9510427  28.3796614
#>  [61]  31.9726126  23.8822570   9.4349371   7.2528177  18.6770037  34.7731315
#>  [67]  41.7298164  25.8310834  86.6112851  52.6902827  28.7531800  22.9506148
#>  [73]  51.8869437  40.0769581  23.3365938   9.4646369  15.1263975   8.9309911
#>  [79]  -7.8001902  -9.4009067   9.8590486  24.9554034  36.6594648   6.6151872
#>  [85]  35.8705825  97.9159300  34.7505020   3.9360583  41.8687838  35.8149797
#>  [91]  28.8288768  -1.9772440   5.5515579   3.5634881 -16.9099012 -16.1928583
#>  [97]  18.5018784  26.4341563  45.4777464  12.8446949  37.2738030  58.4467494
#> [103]  75.7573475  10.2774673  43.5451011  46.2007749  40.9405978   5.8925017
#> [109]  14.8668385  14.0184192  -5.8769301  -4.7633981  24.3481619  36.1679160
#> [115]  41.1631997  38.1304000  55.4032734  55.3350181  35.5344956  63.9844098
#> [121]  68.9856621  48.0698336  33.5428718  36.9270650  39.0559487  25.9151115
#> [127]  14.2233826  11.1424146  -0.6746282  12.0842136  17.8081150  15.5138499
#> [133]  33.9091240  39.9483701  16.5192203  16.8245814 106.4965137  32.9469496
#> [139]  13.8256249  13.1602530  16.6471059   9.5387069 -10.3946313 -12.3258602
#> [145]   8.2219632  20.1129866  30.2377752  13.2101604  39.4244479  52.1914471
#> [151]  37.0839916  14.3061601  52.0550425  86.2299140  36.3621653  10.2820735
#> [157]  23.0310661  21.1906590  -2.1496659  -0.9150761  13.9235054  25.4936434
#> [163]  36.5110377  13.9124510  38.6318014  58.6970725  47.1698326  15.9538492
#> [169]  48.9004962  51.5968814  75.7235283  11.0565840  21.6663900  26.0370826
#> [175]   2.4743166   4.4615118  20.0499963  31.5025875  37.9801227  29.8270028
#> [181]  47.6944968  52.3816955  35.7208893  39.6542353  69.0992804  48.2153639
#> [187]  34.6490192  62.2174191  43.7040983  29.6897866  24.1839983  16.4863442
#> [193]  14.1875010  27.0426229  34.3599636  24.3090550  43.9469710  50.8343256
#> [199]  34.0076940  34.9235098  64.3521495  51.3442693  34.6279026  34.7014921
#> [205]  74.9774857  34.2350093  23.7055064  22.5408562  20.4448333  31.8452090
#> [211]  40.6915988  26.7357114  47.4146255  57.1163205  42.6816015  29.5355797
#> [217]  66.5845244  58.1434081  48.0075628  29.4582908  42.9780720  63.6501606
#> [223]  24.8713236  27.1502721  24.4140625  35.3942183  43.9933794  32.4021211
#> [229]  51.1024750  59.7018750  43.1939761  38.1147555  67.7484478  57.0925197
#> [235]  45.2098887  44.7309850  52.2168856  45.1717284  56.0492251  36.4913157
#> [241]  25.0921360  35.4305381  43.9680359  29.8227571  51.1652837  59.6704380
#> [247]  47.0073572  35.0512983  66.6527891  58.0186237  47.9038379  36.5247501
#> [253]  51.4736513  47.8807649  36.1253678  55.9699358

resid(spflow_results) # vector of residuals

#>   [1] -0.650120852 -0.986046043  2.195392966  0.069549845  0.230373871
#>   [6]  3.285400495  0.808645375 -2.569960347 -1.194311054 -0.995840698
#>  [11]  2.290240790  0.439059189  0.721567854 -0.039074001 -1.212179259
#>  [16]  3.498758864  0.731261924 -3.431156715  1.146159104 -1.186289326
#>  [21] -2.384056721 -0.874415800 -1.955959107 -1.273058069 -0.874957350
#>  [26] -3.260574197  1.795873167  0.297084352 -2.135514294  2.593855542
#>  [31]  1.005087649 -0.457045494  1.167578871  1.476266763  2.299171587
#>  [36]  1.335636935  0.747151287 -0.607012394 -0.973960847 -0.687817598
#>  [41] -1.061743494 -0.286524590 -2.375994940  4.297060958  2.587659656
#>  [46] -2.189453798 -0.683493824 -0.771289667  1.560053666 -0.542158024
#>  [51]  0.350827471 -0.101619098 -0.626030980  2.571007326 -0.678364605
#>  [56]  2.863264084 -3.666525237  1.067597212 -0.028058940  0.426227186
#>  [61]  0.732442401 -1.135248749 -0.634702928 -2.029815220 -2.111603222
#>  [66]  0.362344516  0.524764280 -0.290565915  2.410538599  3.727916044
#>  [71] -0.988433210 -4.920157500  1.863936089 -1.732786533 -1.344645483
#>  [76]  2.116745935 -0.425217210 -2.358457378  0.560401168 -0.021961976
#>  [81] -0.035042088  0.444834360 -1.129104582  1.148631944 -0.832373752
#>  [86]  1.433370149  1.663192067  0.921336404 -0.879378313  1.912171223
#>  [91]  1.483996843  1.070166126  0.596412416 -1.281540494  2.891657024
#>  [96] -1.006482133  4.223491155  3.074015347 -0.949684655 -2.429443759
#> [101] -1.534998393 -0.136477602 -0.096230947 -0.810774242 -1.746595403
#> [106] -0.691531243 -1.963432891 -3.545032193 -0.793711513  1.756462973
#> [111] -1.171859448  1.322957774 -3.245010467  0.023035197  1.112161205
#> [116]  0.420509702 -0.119006675 -1.265757451 -1.724012400 -1.808950606
#> [121] -0.044126959 -1.861631961 -0.995653055 -0.608113173  3.538376922
#> [126] -1.193351589  0.748019839  0.413912979 -1.655405848  0.247939707
#> [131]  3.340702656  0.610163200 -0.037404679 -0.967643500 -3.868846982
#> [136]  2.224225485 -1.565696388  1.444598927  4.190420808 -3.011401611
#> [141]  1.512232902 -0.464888458 -2.623116137 -2.439255552 -3.246264443
#> [146] -0.941825122 -2.798038887  1.190332853  3.877414932 -2.971310949
#> [151]  1.166562102  1.566345521  0.467245234 -1.462820238 -0.552869963
#> [156] -0.521088543  0.966763792 -0.964303530  2.163805534 -0.490114345
#> [161]  2.036773809 -2.064463541 -2.347609438  6.294895005 -0.870046188
#> [166]  0.036814341  0.878603167 -1.087937821  1.261394911  0.271417350
#> [171]  0.096551544  0.160690966  0.073319692  3.983264342 -1.359457241
#> [176] -1.917127602  0.004195526  0.711670335  0.885359392 -0.845678535
#> [181] -2.077648516  2.621570427 -0.836904702 -1.730003245 -0.854849477
#> [186] -0.337739240  2.204395245  0.558184895  1.376515125 -0.846453770
#> [191]  0.527130244 -0.240428232  0.368720296 -1.520122338 -2.363747673
#> [196]  4.117318439  1.400709477 -2.252417544 -1.068021266 -2.421328280
#> [201]  4.433049590  2.558165905 -0.270240650  1.169176060 -0.370962457
#> [206] -0.919226877 -1.423895411 -1.050965523  3.167187605 -0.050980955
#> [211]  0.328610050  0.310213522  2.485517473 -0.923501522 -1.969880295
#> [216]  3.478968165 -1.050265885 -1.654503013 -2.633159868 -2.399978554
#> [221] -1.206927813  1.372611074  2.251709620  1.547838826 -0.652710958
#> [226]  0.416767593 -1.184084485 -1.314183020  2.040483051 -1.794958134
#> [231]  4.112355634  0.238447239  0.929830344 -1.206300985 -0.833570675
#> [236] -2.457833511 -0.216754906  0.933965777  1.004168132 -1.136342813
#> [241] -1.611870266 -0.769841175  3.276963733 -2.181844478 -0.189660416
#> [246]  4.130945573 -0.102815646 -2.366867609 -0.862372297  1.282915682
#> [251] -0.354281047 -0.627106103 -0.689333634  0.233781683  3.544514972
#> [256]  0.312961323

nobs(spflow_results) # number of observations

#> [1] 256

sd_error(spflow_results) # standard deviation of the error term

#> [1] 1.838627

predict(spflow_results) # computation of the in sample predictor

#>   [1]  66.3376429  50.8014499  61.1578140  24.8867148  46.7751209  56.5409569
#>   [7]  39.2213556  19.3684800  47.8924765  39.4692299  30.1327927  12.5518733
#>  [13]  17.4937286  13.7611166  -3.1384868  -4.9896964  39.6004008  75.1356183
#>  [19]  51.8613923  27.9581641  49.2720112  58.8780614  34.5671387  17.7771520
#>  [25]  46.5491519  37.3054794  27.1027159  11.8045778  16.4919493  10.3404487
#>  [31]  -5.4530288  -5.4091486  36.1788546  37.9280988  88.7851439  13.4720536
#>  [37]  42.3959474  55.3871779  38.2001573   9.1388915  39.6996045  31.1902054
#>  [43]  23.2260834   1.9776372   7.9149077   3.7809303 -13.0817487 -12.8476009
#>  [49]  31.9228698  48.2513832  49.8558924  58.2241704  60.8846908  59.6555775
#>  [55]  39.4131208  38.9206190  69.5072855  47.8003554  36.3172222  28.4133805
#>  [61]  32.0052309  23.3150723   9.1617662   6.9376293  18.4539364  34.5795182
#>  [67]  42.0760811  25.8814855  86.6259031  52.9984382  28.2878898  23.1752663
#>  [73]  51.3116529  39.8756903  23.3428555   9.4390820  15.3168790   8.3919011
#>  [79]  -7.5357649  -9.9872519  10.1368600  24.8370713  36.7170887   7.1787753
#>  [85]  36.1444167  98.3620786  34.9204614   3.7692731  42.0761277  35.9386751
#>  [91]  29.1729270  -1.4557405   6.1557931   4.3422111 -16.9098352 -16.2586917
#>  [97]  18.9905075  26.9821591  46.1725906  13.4916958  36.9153519  58.0743666
#> [103]  75.5985343   9.0670117  42.4079800  45.7832285  40.7816929   5.6573502
#> [109]  14.6265045  13.9401182  -6.0164437  -5.0999435  24.2799243  35.4285868
#> [115]  40.5078162  38.1558977  55.0695126  55.3698329  34.7003227  63.8565028
#> [121]  68.5369702  47.8495077  33.3360752  37.2003460  39.4456008  25.6496472
#> [127]  14.3274401  11.0857390  -0.5673441  11.9535943  17.3873391  15.9526411
#> [133]  34.6270919  40.5819410  16.7468527  16.5565371 106.1353643  32.9618042
#> [139]  13.6766559  12.8915394  16.4355673   8.9710474 -10.8301793 -12.9530548
#> [145]   8.3177144  19.0580041  28.9278853  14.2173726  39.4289220  51.9630883
#> [151]  36.3835581  14.5852139  52.2825330  86.3313541  35.9618771  10.4275883
#> [157]  23.2491116  21.5824361  -2.1532969  -0.8872330  13.9756006  26.2889004
#> [163]  36.1788732  13.7646740  39.0549809  58.3782626  47.1596373  16.9804539
#> [169]  49.7020682  51.9659728  75.8875282  10.4183410  21.4661623  26.1457843
#> [175]   2.7238020   4.6069611  19.8669523  31.6693904  38.2023813  29.6357643
#> [181]  47.7897095  52.2450811  36.2119881  39.2671468  69.1556659  48.2831470
#> [187]  34.8157684  61.8445268  44.0111001  29.8532679  24.3150642  16.3252563
#> [193]  13.3017813  27.1965425  34.4007384  24.2889829  44.5480119  50.7838188
#> [199]  33.5788459  35.5722689  65.0967170  51.2766399  34.4692801  33.9726083
#> [205]  75.2211614  34.0541322  24.2287382  21.9809362  20.4198906  32.3316236
#> [211]  41.3024969  27.6456209  48.0640167  56.5055857  42.3489202  29.2399471
#> [217]  66.7582511  57.5944605  47.4864871  29.3319154  43.3535618  64.0448281
#> [223]  25.4493432  27.0318053  24.2945508  34.8938987  44.5375570  32.5361013
#> [229]  50.9114306  60.4513698  42.6875169  37.5909130  67.4398106  57.6224293
#> [235]  45.4632985  44.6228404  51.8497716  44.9570090  56.2458778  36.5697095
#> [241]  25.4508017  35.2144809  43.8325973  29.4262708  51.6133397  59.8346371
#> [247]  47.9071208  34.5862433  66.6191458  58.1326799  48.0049491  36.2510283
#> [253]  51.5192604  48.4204820  36.4983489  56.2708426

# MLE methods
logLik(spflow_results) # value of the likelihood function

#> [1] -529.518

# MLE, OLS and S2SLS methods
varcov(spflow_results) # variance covariance matrix of the estimators

#>                [,1]          [,2]          [,3]        [,4]         [,5]
#>  [1,]  9.123869e-04 -6.793655e-05 -0.0006621455 -0.02107189  0.009951384
#>  [2,] -6.793655e-05  1.139313e-03 -0.0003314255 -0.01946157  0.012729082
#>  [3,] -6.621455e-04 -3.314255e-04  0.0018322455 -0.01979914 -0.012023929
#>  [4,] -2.107189e-02 -1.946157e-02 -0.0197991434  3.67780635 -1.819479137
#>  [5,]  9.951384e-03  1.272908e-02 -0.0120239288 -1.81947914  9.396375955
#>  [6,]  2.202977e-04 -1.995085e-03  0.0005278702  0.02774336 -0.018056531
#>  [7,] -4.823146e-04 -8.703444e-04 -0.0025468380  0.10447065  0.018425444
#>  [8,]  8.360023e-04  2.618291e-04 -0.0004056365 -0.03903218  0.015940208
#>  [9,] -8.967875e-05  1.282269e-03  0.0015980084 -0.12283398  0.047435796
#> [10,] -2.184937e-04 -9.707756e-05  0.0004531018  0.00440309 -0.014488915
#> [11,]  9.792975e-05  4.268554e-05 -0.0002080234  0.06838016 -0.527325356
#> [12,]  3.197900e-03  7.295034e-03  0.0018066746 -0.50977437  0.189826139
#> [13,] -7.400961e-04 -5.870516e-04  0.0007438463  0.02930276 -0.015696280
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,]  0.0002202977 -0.0004823146  0.0008360023 -8.967875e-05 -2.184937e-04
#>  [2,] -0.0019950853 -0.0008703444  0.0002618291  1.282269e-03 -9.707756e-05
#>  [3,]  0.0005278702 -0.0025468380 -0.0004056365  1.598008e-03  4.531018e-04
#>  [4,]  0.0277433619  0.1044706532 -0.0390321832 -1.228340e-01  4.403090e-03
#>  [5,] -0.0180565313  0.0184254443  0.0159402083  4.743580e-02 -1.448891e-02
#>  [6,]  0.0039089691  0.0010896900 -0.0003035879 -2.222584e-03 -3.075792e-04
#>  [7,]  0.0010896900  0.0115475526 -0.0015247438 -5.952458e-03 -1.381565e-04
#>  [8,] -0.0003035879 -0.0015247438  0.0013406115  4.656331e-04 -6.251463e-04
#>  [9,] -0.0022225837 -0.0059524579  0.0004656331  6.781953e-03  6.315897e-04
#> [10,] -0.0003075792 -0.0001381565 -0.0006251463  6.315897e-04  6.626335e-03
#> [11,]  0.0002607544 -0.0021221303  0.0004001936 -2.490730e-03 -5.501756e-03
#> [12,] -0.0115261673 -0.0249219880  0.0074825540  1.901875e-02 -1.553531e-03
#> [13,]  0.0009413649  0.0009273644 -0.0008638438 -6.635535e-04  2.395589e-04
#>               [,11]         [,12]         [,13]
#>  [1,]  9.792975e-05  0.0031979000 -0.0007400961
#>  [2,]  4.268554e-05  0.0072950340 -0.0005870516
#>  [3,] -2.080234e-04  0.0018066746  0.0007438463
#>  [4,]  6.838016e-02 -0.5097743698  0.0293027620
#>  [5,] -5.273254e-01  0.1898261391 -0.0156962804
#>  [6,]  2.607544e-04 -0.0115261673  0.0009413649
#>  [7,] -2.122130e-03 -0.0249219880  0.0009273644
#>  [8,]  4.001936e-04  0.0074825540 -0.0008638438
#>  [9,] -2.490730e-03  0.0190187536 -0.0006635535
#> [10,] -5.501756e-03 -0.0015535313  0.0002395589
#> [11,]  3.747948e-02 -0.0003704992 -0.0001068666
#> [12,] -3.704992e-04  0.1150842643 -0.0069068502
#> [13,] -1.068666e-04 -0.0069068502  0.0902460998

# MCMC methods
if (FALSE) {
mcmc_results(spflow_results) # parameter values during the mcmc sampling
}