Create a spflow_network_pair-class()

Create a spflow_network_pair-class()

spflow_network_pair(
  id_orig_net,
  id_dest_net,
  id_net_pair = paste0(id_orig_net, "_", id_dest_net),
  pair_data = NULL,
  orig_key_column,
  dest_key_column
)

Arguments

id_orig_net

A character that serves as identifier for the origin network

id_dest_net

A character that serves as identifier for the destination network

id_net_pair

A character that as identifier for network_pair

pair_data

A data.frame containing information on the origin-destination pairs

orig_key_column

A character indicating the name of the column containing the identifiers of the origins

dest_key_column

A character indicating the name of the column containing the identifiers of the destinations

Value

A spflow_network_pair-class()

Examples

pair_frame <- data.frame(
  ORIG_ID_STATE = rep(germany_grid$ID_STATE, times = 16),
  DEST_ID_STATE = rep(germany_grid$ID_STATE, each = 16))
spflow_network_pair("ge","ge","ge_ge",pair_frame,"ORIG_ID_STATE","DEST_ID_STATE")
#> Spatial network pair with id: ge_ge
#> --------------------------------------------------
#> Origin network id: ge (with 16 nodes)
#> Destination network id: ge (with 16 nodes)
#> Number of pairs: 256
#> Completeness of pairs: 100.00% (256/256)
#> 
#> Data on node-pairs:
#>     DEST_ID_STATE ORIG_ID_STATE
#> 1              SH            SH
#> 17             HH            SH
#> 33             MV            SH
#> 49             NW            SH
#> 65             HB            SH
#> 81             BB            SH
#> ---           ---           ---
#> 176            SN            BY
#> 192            SL            BY
#> 208            HE            BY
#> 224            TH            BY
#> 240            BW            BY
#> 256            BY            BY
#>