Spatial econometric interaction models in R

The spflow package provides function to estimate models of origin-destination flows with spatial autocorrelation. The implementation exploits the efficiency gains of the matrix formulation that avoids duplication of the data that describes the origins and the destinations. It offers new data structures and a convenient formula interface that allows to estimate the models with maximal efficiency and minimal effort for the user. The estimators are detailed in Dargel (2021) .

Network data

The package provides a new class for representing network data. These classes help to exploit the relational structure of the data and ensure that all origins and destinations are correctly identified.

Model estimation

The spflow() function is the main function for estimation and can be used out of the box using the default configuration, which corresponds to the most general model possible. Fine grained control over the estimation procedure is given via the spflow_control() argument through which the user can change, for example, the estimation method or the way in which the covariates should be used.

References

Dargel L (2021). “Revisiting Estimation Methods for Spatial Econometric Interaction Models.” Toulouse School of Economics.

See also

Author

Maintainer: Lukas Dargel lukas.dargel@mailbox.org (ORCID)

Authors:

Thibault Laurent thibault.laurent@tse-fr.eu (ORCID)

Other contributors:

Christine Thomas christine.thomas@tse-fr.eu [thesis advisor]

Paula Margaretic [contributor]

Gabriel Watkinson [contributor]

Roger Bivand [contributor]